+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - removed strict type safety from ExtendedEnum
 - added angular histogram
 - fixed XYZ axes being mirrored
+- fixed camera tracking and camera velocity
+- added camera orbit to camera jobs
 
 ## 2021-07-06
 - explicitly setting path type for all path entries